+++ R/shiny_vector_filter.R
@@ -106,7 +106,8 @@ shiny_vector_filter <- function(data, inputId, global = FALSE) {
 #' @keywords internal
 shiny_vector_filter.default <- function(data, inputId, ...) {
   function(input, output, session, x = shiny::reactive(NULL), 
-      filter_na = shiny::reactive(FALSE), verbose = FALSE) { 
+      filter_na = shiny::reactive(FALSE), filter_fn = NULL, verbose = FALSE,
+      erase_filters = shiny::reactive(0)) { 
     
     module_return <- shiny::reactiveValues(code = FALSE, mask = FALSE)
     module_return$code <- shiny::reactive(FALSE)
@@ -125,7 +126,7 @@ shiny_vector_filter.default <- function(data, inputId, ...) {
 #'
 #' @return a pillar formatted class name
 #'
-#' @importFrom pillar new_pillar_type
+#' @importFrom pillar type_sum
 #' @keywords internal
 #' 
 get_dataFilter_class <- function(obj) {
@@ -139,5 +140,6 @@ get_dataFilter_class <- function(obj) {
   
   if (!length(vf_class)) return("unk")
   class(obj) <- vf_class
-  pillar::new_pillar_type(obj)[[1]][1]
+  type <- pillar::type_sum(obj)
+  if (length(type) == 0L) "" else type
 }